Fix serialization of the "key" property
The introduction of a payload envelope to the update notification event
issued by champ-core caused the incorrect serialization of the "key"
property as a JSON object instead of a String.
This caused Spike to fail when parsing the notification event from
Champ as it expects "key" to be a String.

A previous fix attempt incorrectly addressed this problem. As the issue
originated from the inconsistent use of both Jackson and Gson JSON
libraries with the envelope using Gson but the contained objects
using Jackson annotations. This fix primarily involves updates to
champ-core to ensure that the envelope generation uses Jackson
instead of Gson and respects the original Jackson serialization of
the "key" property as a String.

Change-Id: I2a12732c9ff3970c3db9de5f0039304e68cb3556
Issue-ID: AAI-1243
Signed-off-by: Michael Arrastia <MArrasti@amdocs.com>